+++ R/pr.R
@@ -635,17 +635,9 @@ pr_clean <- function(
     tryCatch(
       gert::git_branch_delete(pr_local_branch, repo = repo),
       libgit2_error = function(e) {
-        if (identical(Sys.getenv("USER"), "jenny")) {
-          write_to <- glue("~/rrr/usethis/pr-finish-oops-{project_name()}.rds")
-          saveRDS(e, write_to)
-          ui_bullets(c(
-            "!" = "Wrote an error from {.fun git_branch_delete} to {.path {pth(write_to, base = NA)}}!"
-          ))
-        }
         # The expected error doesn't have a distinctive class, so we have to
         # detect it based on the message.
-        # If we got an unexpected libgit2 error, rethrow.
-
+        # If we get an unexpected libgit2 error, rethrow.
         if (
           !grepl(
             "could not find key 'branch[.].+[.](vscode-merge-base|github-pr-owner-number)' to delete",