+++ R/proj.R
@@ -14,18 +14,25 @@ proj_set_ <- function(path) {
 #' Most `use_*()` functions act on the **active project**. If it is
 #' unset, usethis uses [rprojroot](https://rprojroot.r-lib.org) to
 #' find the project root of the current working directory. It establishes the
-#' project root by looking for a `.here` file, an RStudio Project, a package
-#' `DESCRIPTION`, Git infrastructure, a `remake.yml` file, or a `.projectile`
-#' file. It then stores the active project for use for the remainder of the
+#' project root by looking for signs such as:
+#' * a `.here` file
+#' * an RStudio Project, i.e. a `.Rproj` file
+#' * an R package, i.e. a `DESCRIPTION` file
+#' * a Git repository
+#' * a Positron or VS Code workspace, i.e. a `.vscode/settings.json` file
+#' * a Quarto project, i.e. a `_quarto.yml` file
+#' * an renv project, i.e. a `renv.lock` file
+#'
+#' usethis then stores the active project for use for the remainder of the
 #' session.
 #'
 #' In general, end user scripts should not contain direct calls to
 #' `usethis::proj_*()` utility functions. They are internal functions that are
 #' exported for occasional interactive use or use in packages that extend
-#' usethis. End user code should call functions in
-#' [rprojroot](https://rprojroot.r-lib.org) or its simpler companion,
-#' [here](https://here.r-lib.org), to programmatically detect a project and
-#' build paths within it.
+#' usethis. End user code should call `here::here()` or other functions from
+#' the [here](https://here.r-lib.org) or
+#' [rprojroot](https://rprojroot.r-lib.org) packages to programmatically
+#' detect a project and build paths within it.
 #'
 #' If you are puzzled why a path (usually the current working directory) does
 #' *not* appear to be inside project, it can be helpful to call
@@ -190,6 +197,10 @@ proj_crit <- function() {
     rprojroot::is_rstudio_project |
     rprojroot::is_r_package |
     rprojroot::is_git_root |
+    # use rprojroot::is_vscode_project at some point in the future
+    rprojroot::has_file(".vscode/settings.json") |
+    rprojroot::is_quarto_project |
+    rprojroot::is_renv_project |
     rprojroot::is_remake_project |
     rprojroot::is_projectile_project
 }
@@ -279,8 +290,10 @@ proj_activate <- function(path) {
   path <- user_path_prep(path)
 
   if (rstudio_available() && rstudioapi::hasFun("openProject")) {
+    # TODO: Perhaps in future this message can be specialized for RStudio vs.
+    # Positron. For now, I've just made it more generic to work better for both.
     ui_bullets(c(
-      "v" = "Opening {.path {pth(path, base = NA)}} in new RStudio session."
+      "v" = "Opening {.path {pth(path, base = NA)}} in a new session."
     ))
     rstudioapi::openProject(path, newSession = TRUE)
     invisible(FALSE)