+++ tests/testthat/test-usethis-deprecated.R
@@ -4,3 +4,7 @@ test_that("use_tidy_eval() is deprecated", {
   pkg <- create_local_package()
   expect_snapshot(use_tidy_eval(), error = TRUE)
 })
+
+test_that("use_tidy_style() is deprecated", {
+  expect_snapshot(use_tidy_style())
+})